air_mobile_app/raspberry_pi/requirements.txt
pyaudio==0.2.13
numpy==1.24.3
requests==2.31.0
openai-whisper==20231117
wave==0.0.2 